責任を別のクラスに転換する
